// 2018/tinglewapp-develop/src/models/social-event.interface.ts
export interface SocialEvent {
  id?: string;
  distance?: number;
  eventTitle: string;
  eventDescription: string;
  eventStartDate: string;
  eventEndDate: string;
  eventStartTime: string;
  eventEndTime: string;
  eventTicketType: string;
  eventTicketPrice: string;
  eventRefundPolicy: string;
  eventAddress: string;
  eventType: string;
  eventAccess: string;
  lat: string;
  lng: string;
  uid: string;
}